Update Group SPD endpoint request tests
