+++ explainability-service/src/test/java/org/kie/trustyai/service/endpoints/metrics/fairness/group/GroupStatisticalParityDifferenceRequestsEndpointMemoryTest.java
@@ -1,0 +1,56 @@
+package org.kie.trustyai.service.endpoints.metrics.fairness.group;
+
+import java.util.Map;
+import java.util.UUID;
+
+import org.jboss.resteasy.reactive.RestResponse;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.kie.trustyai.explainability.model.Dataframe;
+import org.kie.trustyai.service.config.ServiceConfig;
+import org.kie.trustyai.service.endpoints.metrics.MetricsEndpointTestProfile;
+import org.kie.trustyai.service.endpoints.metrics.RequestPayloadGenerator;
+import org.kie.trustyai.service.mocks.MockDatasource;
+import org.kie.trustyai.service.mocks.MockMemoryStorage;
+import org.kie.trustyai.service.mocks.MockPrometheusScheduler;
+import org.kie.trustyai.service.payloads.BaseScheduledResponse;
+import org.kie.trustyai.service.payloads.metrics.BaseMetricRequest;
+import org.kie.trustyai.service.payloads.metrics.fairness.group.GroupMetricRequest;
+import org.kie.trustyai.service.payloads.scheduler.ScheduleList;
+
+import io.quarkus.test.common.http.TestHTTPEndpoint;
+import io.quarkus.test.junit.QuarkusTest;
+import io.quarkus.test.junit.TestProfile;
+import io.restassured.http.ContentType;
+
+import jakarta.enterprise.inject.Instance;
+import jakarta.inject.Inject;
+import jakarta.ws.rs.core.Response;
+import org.kie.trustyai.service.profiles.MemoryTestProfile;
+
+import static io.restassured.RestAssured.given;
+import static org.hamcrest.Matchers.containsString;
+import static org.junit.jupiter.api.Assertions.*;
+
+@QuarkusTest
+@TestProfile(MemoryTestProfile.class)
+@TestHTTPEndpoint(GroupStatisticalParityDifferenceEndpoint.class)
+class GroupStatisticalParityDifferenceRequestsEndpointMemoryTest extends GroupStatisticalParityDifferenceRequestsEndpointBaseTest {
+
+    @Inject
+    Instance<MockMemoryStorage> storage;
+
+    @BeforeEach
+    void populateStorage() {
+        // Empty mock storage
+        storage.get().emptyStorage();
+        // Clear any requests between tests
+        scheduler.get().getAllRequestsFlat().clear();
+        final Dataframe dataframe = datasource.get().generateRandomDataframe(1000);
+        datasource.get().saveDataframe(dataframe, MODEL_ID);
+        datasource.get().saveMetadata(datasource.get().createMetadata(dataframe), MODEL_ID);
+    }
+
+
+}
